+++ app/src/main/java/br/edu/ifsp/scl/controlefinanceiro/model/Conta.kt
@@ -2,7 +2,7 @@ package br.edu.ifsp.scl.controlefinanceiro.model
 
 // Valores padrões são Strings vazias
 data class Conta(
-    var idConta: Int = 0,
+    var idConta: Long = 0L,
     var descricao: String ="",
     var saldo: Float = 0F
