+++ app/src/main/java/br/edu/ifsp/scl/controlefinanceiro/model/Transacao.kt
@@ -3,7 +3,7 @@ package br.edu.ifsp.scl.controlefinanceiro.model
 data class Transacao(
     var idTransacao: Int = 0,
     var descricao: String = "",
-    var idConta: Int = 0,
+    var idConta: Long = 0L,
     var valor: Float = 0F,
     var natureza: String = "",
     var tipo: String = ""
